[Bug #161893968] Refactor views init methods in views
